Split out libs/ballot-iterpreter 'interpret_*' tests
