+++ libs/ballot-interpreter/e2e/BUILD.bazel
@@ -1,0 +1,39 @@
+load("//tools/ts_build:lint_test.bzl", "lint_test")
+load("//tools/ts_build:ts_tests.bzl", "ts_tests")
+
+ts_tests(
+    name = "tests",
+    size = "enormous",
+    data = ["//libs/ballot-interpreter/test:fixtures_js"],
+    shard_count = 4,
+    deps = [
+        "//:node_modules/@types/jest",
+        "//:node_modules/@types/node",
+        "//:node_modules/@types/testing-library__jest-dom",
+        "//:node_modules/canvas",
+        "//libs/ballot-encoder/src",
+        "//libs/ballot-interpreter/src",
+        "//libs/ballot-interpreter/test",
+        "//libs/basics/assert",
+        "//libs/basics/collections",
+        "//libs/basics/iterators",
+        "//libs/basics/types",
+        "//libs/bmd-ballot-fixtures/src",
+        "//libs/fixtures/src/data/electionFamousNames2021",
+        "//libs/fixtures/src/data/electionGeneral",
+        "//libs/fixtures/src/data/electionGridLayoutNewHampshireTestBallot",
+        "//libs/fixtures/src/data/electionPrimaryPrecinctSplits",
+        "//libs/fixtures/src/data/sample-ballot-images",
+        "//libs/fs/src",
+        "//libs/hmpb/src",
+        "//libs/image-test-utils",
+        "//libs/image-utils/src",
+        "//libs/test-utils/src",
+        "//libs/types/election-parsing",
+        "//libs/types/elections",
+        "//libs/types/scanning",
+        "//libs/utils/src",
+    ],
+)
+
+lint_test(name = "lint")
