+++ libs/ballot-interpreter/e2e/__image_snapshots__/interpret-vx-hmpb-test-ts-hmpb-general-election-letter-paper-language-en-marked-ballot-interpretation-1-snap.png
