+++ libs/ballot-interpreter/src/BUILD.bazel
@@ -41,10 +41,8 @@ ts_library(
 ts_tests(
     name = "tests",
     size = "enormous",
-    timeout = "moderate",
     data = ["//libs/ballot-interpreter/test:fixtures_js"],
     jest_environment = "node",  # (Defaults to jsdom for any package that imports React.)
-    shard_count = 4,
     deps = [
         "//:node_modules/@types/jest",
         "//:node_modules/@types/node",
@@ -61,20 +59,15 @@ ts_tests(
         "//libs/basics/collections",
         "//libs/basics/iterators",
         "//libs/basics/result",
-        "//libs/basics/types",
         "//libs/bmd-ballot-fixtures/src",
         "//libs/fixtures/src/data/electionFamousNames2021",
         "//libs/fixtures/src/data/electionGeneral",
         "//libs/fixtures/src/data/electionGridLayoutNewHampshireTestBallot",
-        "//libs/fixtures/src/data/electionPrimaryPrecinctSplits",
         "//libs/fixtures/src/data/electionTwoPartyPrimary",
         "//libs/fixtures/src/data/sample-ballot-images",
-        "//libs/fs/src",
-        "//libs/hmpb/src",
         "//libs/image-test-utils",
         "//libs/image-utils/src",
         "//libs/test-utils/src",
-        "//libs/types/election-parsing",
         "//libs/types/elections",
         "//libs/types/scanning",
         "//libs/utils/src",
